+++ src/components/App/components/StatusMessage.vue
@@ -3,36 +3,51 @@
     <div
       v-if="isShown"
       :class="`status-message status-message--${messageType}`"
+      @mouseenter="stopProgressBar"
+      @mouseleave="startProgressBar"
     >
-      <div class="status-message__icon-wrp">
-        <i
-          class="mdi status-message__icon"
-          :class="[`status-message__icon--${messageType}`, messageIconClass]"
+      <div class="status-message__body">
+        <div class="status-message__icon-wrp">
+          <i
+            class="mdi status-message__icon"
+            :class="[`status-message__icon--${messageType}`, messageIconClass]"
+          />
+        </div>
+
+        <div class="status-message__payload">
+          <h4 class="status-message__title">
+            {{ messageTitle }}
+          </h4>
+
+          <p class="status-message__text">
+            {{ message }}
+          </p>
+        </div>
+
+        <button
+          @click="isShown = false"
+          class="status-message__close-btn"
         />
       </div>
-
-      <div class="status-message__payload">
-        <h4 class="status-message__title">
-          {{ messageTitle }}
-        </h4>
-
-        <p class="status-message__text">
-          {{ message }}
-        </p>
+      <div class="status-message__progress-bar">
+        <span
+          class="status-message__progress-bar-percentage"
+          :style="{ 'width': getInvertedProgressPercents + '%' }"
+          :class="`status-message__progress-bar-percentage--${messageType}`"
+        />
       </div>
-
-      <button
-        @click="isShown = false"
-        class="status-message__close-btn"
-      />
     </div>
   </transition>
 </template>
 
 <script>
 import { Bus } from '@/utils/bus'
+import { MathUtil } from '@/utils/math.util'
 
 const CLOSE_TIMEOUT_MS = 10000
+const ONE_HUNDRED_PERCENTS = 100
+const MAX_ANIMATION_PROGRESS = 1
+const MIN_ANIMATION_PROGRESS = 0
 const MESSAGE_TYPES = Object.freeze({
   warning: 'warning',
   success: 'success',
@@ -48,7 +63,11 @@ export default {
     messageType: '',
     messageArgs: {},
     isShown: false,
-    timeoutId: -1,
+    animationFrame: -1,
+    progressBar: {
+      progress: 0,
+      paused: false,
+    },
   }),
 
   computed: {
@@ -81,6 +100,19 @@ export default {
           return ''
       }
     },
+
+    getInvertedProgressPercents () {
+      const animationProgressPercents = MathUtil.multiply(
+        this.progressBar.progress,
+        ONE_HUNDRED_PERCENTS
+      )
+
+      const animationProgressPercentsInverted = MathUtil.subtract(
+        ONE_HUNDRED_PERCENTS,
+        animationProgressPercents
+      )
+      return animationProgressPercentsInverted
+    },
   },
 
   created () {
@@ -94,19 +126,67 @@ export default {
       this.show(MESSAGE_TYPES.info, payload))
   },
 
+  destroyed () {
+    cancelAnimationFrame(this.animationFrame)
+  },
+
   methods: {
     show (messageType, payload) {
       this.messageType = messageType
       this.message = payload
+      cancelAnimationFrame(this.animationFrame)
+      this.progressBar.progress = MIN_ANIMATION_PROGRESS
       this.isShown = true
+      this.startAnimationTimeout()
+    },
 
-      if (this.timeoutId >= 0) {
-        window.clearTimeout(this.timeoutId)
+    startAnimationTimeout (currentAnimationTime = 0) {
+      const animationStartTime = performance.now()
+      const calculateAnimationProgress = () => {
+        this.animationFrame = requestAnimationFrame((timestamp) => {
+          if (this.progressBar.paused) {
+            cancelAnimationFrame(this.animationFrame)
+            return
+          }
+
+          const animationTime = MathUtil.subtract(
+            currentAnimationTime,
+            animationStartTime
+          )
+          const animationRuntime = MathUtil.add(
+            timestamp,
+            animationTime
+          )
+          const animationProgress = MathUtil.divide(
+            animationRuntime,
+            CLOSE_TIMEOUT_MS
+          )
+
+          if (animationRuntime < CLOSE_TIMEOUT_MS) {
+            this.progressBar.progress = animationProgress
+            calculateAnimationProgress()
+          } else {
+            this.progressBar.progress = MAX_ANIMATION_PROGRESS
+            cancelAnimationFrame(this.animationFrame)
+            this.isShown = false
+          }
+        })
       }
+      calculateAnimationProgress()
+    },
 
-      this.timeoutId = window.setTimeout(_ => {
-        this.isShown = false
-      }, CLOSE_TIMEOUT_MS)
+    stopProgressBar () {
+      this.progressBar.paused = true
+    },
+
+    startProgressBar () {
+      this.progressBar.paused = false
+      const currentAnimationTime = MathUtil.multiply(
+        CLOSE_TIMEOUT_MS,
+        this.progressBar.progress
+      )
+
+      this.startAnimationTimeout(currentAnimationTime)
     },
   },
 }
@@ -126,6 +206,7 @@ $icon-padding: 2.4rem;
   max-width: 42rem;
   min-width: 32rem;
   display: flex;
+  flex-direction: column;
   box-shadow: 0 0.4rem 1rem 0 rgba(0, 0, 0, 0.15);
 
   &--warning {
@@ -142,6 +223,43 @@ $icon-padding: 2.4rem;
 
   &--info {
     background-color: $color-content-bg;
+  }
+}
+
+.status-message__body {
+  display: flex;
+}
+
+.status-message__progress-bar {
+  position: absolute;
+  width: 100%;
+  height: 0.5rem;
+  bottom: 0;
+}
+
+.status-message__progress-bar-percentage {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 0.5rem;
+  width: 100%;
+  background-color: $color-success;
+  max-width: 100%;
+
+  &--success {
+    background-color: $color-success;
+  }
+
+  &--error {
+    background-color: $color-text-inverse;
+  }
+
+  &--info {
+    background-color: $color-info;
+  }
+
+  &--warning {
+    background-color: $color-text-inverse;
   }
 }