+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Added
+- Autoclose indicator to status-messages
+
 ## [1.10.0-rc.1] - 2019-07-16
 - Now using @tokend/js-sdk@1.9.0-rc.1