+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Fixed
+- Resolved the issue when admin can't create asset
+
 ## [1.10.0-rc.2] - 2019-08-02
 #### Added
 - Autoclose indicator to status-messages